requirements.txt
Pillow==5.4.1
imgaug==0.2.8
